+++ internal/storage/storage.go
@@ -11,7 +11,7 @@ type Storage interface {
 	Upload(ctx context.Context, file []byte, metadata *CheckinMetadata) error
 	Download(ctx context.Context, fileName string) ([]byte, error)
 	CheckinExists(ctx context.Context, checkinID, createdAt string) (bool, error)
-	GetLatestCheckinID(ctx context.Context) (int, error)
+	GetLatestCheckinID(ctx context.Context) (uint64, error)
 	UpdateLatestCheckinID(ctx context.Context, checkin untappd.Checkin) error
 }
